Lift duplicated positivity propagation into core/
The scalar and t2s positivity_propagation helpers were nearly
identical: same `view` struct, same `mark_*` and `propagate_*`
functions, same assertion macro. Only `read()` was truly
domain-specific (scalar inspects scalar_zero/one/constant/negative
for the numeric-constant check; t2s forwards through
tensor_to_scalar_scalar_wrapper).

New shared header `core/positivity_propagation.h` holds:
 - view struct + at_least_nonneg_tag/at_least_nonpos_tag
 - NUMSIM_CAS_POSITIVITY_ASSERT_NO_CONTRADICTION macro
 - mark_positive/negative/nonnegative/nonpositive (templated on Expr)
 - propagate_mul_from_views/neg_from_view/pow_from_views
   (also templated)

Each domain header now holds only its own read() function:
 - scalar_positivity_propagation.h: ~50 lines (was ~166)
 - tensor_to_scalar_positivity_propagation.h: ~80 lines (was ~245)

Backward-compat namespace aliases (scalar_detail::positivity::*
and tensor_to_scalar_detail::positivity::*) re-export the shared
helpers under their old names so the operator-file call sites
do not need to change.

Net: −89 lines, single source of truth for the rules, easier to
keep scalar/t2s aligned when adding new rules (e.g. when #306 and
#307 land sign-aware and add/sub rules respectively).

Local: 1545 main + 783 fuzz tests pass.
