+++ include/numsim_cas/scalar/scalar_positivity_propagation.h
@@ -1,43 +1,26 @@
 #ifndef SCALAR_POSITIVITY_PROPAGATION_H
 #define SCALAR_POSITIVITY_PROPAGATION_H
 
-#include <cassert>
-
-#include <numsim_cas/core/assumptions.h>
-#include <numsim_cas/core/expression_holder.h>
+#include <numsim_cas/core/positivity_propagation.h>
 #include <numsim_cas/scalar/scalar_constant.h>
 #include <numsim_cas/scalar/scalar_expression.h>
 #include <numsim_cas/scalar/scalar_negative.h>
 #include <numsim_cas/scalar/scalar_one.h>
 #include <numsim_cas/scalar/scalar_zero.h>
 
-// #305 — positivity-tag propagation for scalar operators (mul, neg, pow).
-//
-// Direct parallel of `tensor_to_scalar_positivity_propagation.h` (#260).
-// Same rule table, same aliasing-mutation contract, same defense-in-depth
-// asserts. The only structural difference: scalar has no wrapper-bridge
-// node, so `read()` doesn't need the scalar_wrapper-forwarding branch.
-//
-// When #305 and #260 are both green, the obvious refactor is to lift
-// both into a templated `<Domain>` header. Deferred to #262 (cross-
-// domain unification) to keep this PR focused.
+// Scalar adapter for the domain-agnostic positivity propagation
+// (see `core/positivity_propagation.h`). Only the `read()` function
+// is domain-specific: it inspects scalar-domain leaf node types
+// (scalar_zero/one/constant/negative) for the "is real-by-
+// construction numeric constant" check. Everything else
+// (mark_*, propagate_*, view) is shared.
 
-namespace numsim::cas::scalar_detail::positivity {
+namespace numsim::cas::positivity {
 
-// Tag-state view. NOTE: these are tag-manager predicates, not truth
-// predicates — a caller writing incoherent tags defeats inference.
-struct view {
-  bool positive;
-  bool nonnegative;
-  bool nonpositive;
-  bool negative;
-  bool nonzero;
-  bool real;
-
-  bool at_least_nonneg_tag() const { return positive || nonnegative; }
-  bool at_least_nonpos_tag() const { return negative || nonpositive; }
-};
-
+// Read sign tags from a scalar expression's assumption manager.
+// Promotes the "concrete numeric constant" check to `real` so
+// `pow(x, integer)` exponents fire the real-exponent guard without
+// waiting on #261 (constants pre-annotation).
 inline view read(expression_holder<scalar_expression> const &e) {
   // Invalid-holder guard: an invalid holder would null-deref through
   // numeric_assumption_manager::contains. This is reached from any
@@ -47,10 +30,9 @@ inline view read(expression_holder<scalar_expression> const &e) {
   //     constructs an invalid lhs that the *= safety net would
   //     normally handle silently, but our read() bypasses it.
   //   * diff visitor accumulator state — was the source of the
-  //     FuzzyScalarDiff seed 35 crash that surfaced this latent UB.
-  //     scalar_differentiation now identity-inits its accumulators
-  //     (scalar_zero / scalar_one), so the diff-internal source is
-  //     closed. The guard remains for the user-code path and any
+  //     FuzzyScalarDiff seed 35 crash. scalar_differentiation now
+  //     identity-inits its accumulators so the diff-internal source
+  //     is closed. The guard remains for the user-code path and any
   //     future visitor that doesn't follow the identity-init
   //     pattern.
   if (!e.is_valid())
@@ -65,13 +47,10 @@ inline view read(expression_holder<scalar_expression> const &e) {
              a.contains(numsim::cas::integer{}) ||
              a.contains(numsim::cas::rational{}) ||
              a.contains(numsim::cas::irrational{})};
-  // Concrete numeric constants are real by construction. See the t2s
-  // header's same-shape note (#261 tracks the broader fix of pre-
-  // annotating constant nodes; this branch avoids waiting on it).
-  // Open-coded structural check mirroring domain_traits::try_numeric
-  // (we can't include scalar_domain_traits.h here due to the
-  // scalar_std.h cycle — domain_traits includes scalar_all → scalar_std,
-  // and scalar_std includes this header).
+  // Concrete numeric constants are real by construction. Open-coded
+  // structural check (instead of domain_traits::try_numeric) to
+  // sidestep the scalar_domain_traits → scalar_all → scalar_std
+  // include cycle.
   if (!v.real) {
     if (is_same<scalar_zero>(e) || is_same<scalar_one>(e) ||
         is_same<scalar_constant>(e))
@@ -86,99 +65,22 @@ inline view read(expression_holder<scalar_expression> const &e) {
   return v;
 }
 
-// Defense-in-depth: catches a future rule that would insert a tag
-// contradicting the operand's existing state. Aliasing makes this
-// critical — if `result` shares its node with an operand (via a fold
-// returning the operand directly), the insertion mutates the operand.
-#ifndef NDEBUG
-#define NUMSIM_CAS_SCALAR_PROP_ASSERT_NO_CONTRADICTION(manager, banned_tag)    \
-  do {                                                                         \
-    assert(!(manager).contains(banned_tag) &&                                  \
-           "scalar positivity_propagation: rule would set a tag "              \
-           "contradicting the operand's existing state");                      \
-  } while (0)
-#else
-#define NUMSIM_CAS_SCALAR_PROP_ASSERT_NO_CONTRADICTION(manager, banned_tag)    \
-  ((void)0)
-#endif
+} // namespace numsim::cas::positivity
 
-inline void mark_positive(expression_holder<scalar_expression> const &e) {
-  auto &a = e.data()->assumptions();
-  NUMSIM_CAS_SCALAR_PROP_ASSERT_NO_CONTRADICTION(a, numsim::cas::negative{});
-  NUMSIM_CAS_SCALAR_PROP_ASSERT_NO_CONTRADICTION(a, numsim::cas::nonpositive{});
-  a.insert(numsim::cas::positive{});
-  a.insert(numsim::cas::nonnegative{});
-  a.insert(numsim::cas::nonzero{});
-  a.insert(numsim::cas::real_tag{});
-  a.set_inferred();
-}
-inline void mark_negative(expression_holder<scalar_expression> const &e) {
-  auto &a = e.data()->assumptions();
-  NUMSIM_CAS_SCALAR_PROP_ASSERT_NO_CONTRADICTION(a, numsim::cas::positive{});
-  NUMSIM_CAS_SCALAR_PROP_ASSERT_NO_CONTRADICTION(a, numsim::cas::nonnegative{});
-  a.insert(numsim::cas::negative{});
-  a.insert(numsim::cas::nonpositive{});
-  a.insert(numsim::cas::nonzero{});
-  a.insert(numsim::cas::real_tag{});
-  a.set_inferred();
-}
-inline void mark_nonnegative(expression_holder<scalar_expression> const &e) {
-  auto &a = e.data()->assumptions();
-  NUMSIM_CAS_SCALAR_PROP_ASSERT_NO_CONTRADICTION(a, numsim::cas::negative{});
-  a.insert(numsim::cas::nonnegative{});
-  a.insert(numsim::cas::real_tag{});
-  a.set_inferred();
-}
-inline void mark_nonpositive(expression_holder<scalar_expression> const &e) {
-  auto &a = e.data()->assumptions();
-  NUMSIM_CAS_SCALAR_PROP_ASSERT_NO_CONTRADICTION(a, numsim::cas::positive{});
-  a.insert(numsim::cas::nonpositive{});
-  a.insert(numsim::cas::real_tag{});
-  a.set_inferred();
-}
-
-// Mul: pos·pos → pos; (≥nonneg)·(≥nonneg) → nonneg. ORDER MATTERS.
-inline void
-propagate_mul_from_views(view lhs, view rhs,
-                         expression_holder<scalar_expression> const &result) {
-  if (lhs.positive && rhs.positive) {
-    mark_positive(result);
-  } else if (lhs.at_least_nonneg_tag() && rhs.at_least_nonneg_tag()) {
-    mark_nonnegative(result);
-  }
-}
-
-// Neg: flip sign, preserve magnitude class. The -(-x) → x fold lives
-// in the cpp file and bypasses this path. The -neg / -nonpos branches
-// are kept for future-symmetry with sign-aware mul (out of scope).
-inline void
-propagate_neg_from_view(view operand,
-                        expression_holder<scalar_expression> const &result) {
-  if (operand.positive) {
-    mark_negative(result);
-  } else if (operand.at_least_nonneg_tag()) {
-    mark_nonpositive(result);
-  } else if (operand.negative) {
-    mark_positive(result);
-  } else if (operand.at_least_nonpos_tag()) {
-    mark_nonnegative(result);
-  }
-}
-
-// Pow: pow(pos, real) → pos; pow(≥nonneg, ≥nonneg) → nonneg.
-// ORDER MATTERS.
-inline void
-propagate_pow_from_views(view base, view exponent,
-                         expression_holder<scalar_expression> const &result) {
-  if (base.positive && exponent.real) {
-    mark_positive(result);
-    return;
-  }
-  if (base.at_least_nonneg_tag() && exponent.at_least_nonneg_tag()) {
-    mark_nonnegative(result);
-  }
-}
-
+// Backward-compat namespace alias: existing scalar-domain call sites
+// reference `scalar_detail::positivity::*`. Re-export the shared
+// helpers + the scalar read() under that name so the operator-file
+// instrumentation doesn't need to change.
+namespace numsim::cas::scalar_detail::positivity {
+using numsim::cas::positivity::mark_negative;
+using numsim::cas::positivity::mark_nonnegative;
+using numsim::cas::positivity::mark_nonpositive;
+using numsim::cas::positivity::mark_positive;
+using numsim::cas::positivity::propagate_mul_from_views;
+using numsim::cas::positivity::propagate_neg_from_view;
+using numsim::cas::positivity::propagate_pow_from_views;
+using numsim::cas::positivity::read;
+using numsim::cas::positivity::view;
 } // namespace numsim::cas::scalar_detail::positivity
 
 #endif // SCALAR_POSITIVITY_PROPAGATION_H
